# set pachd's $GOMEMLIMIT from k8s requests or limits