--- a/etc/helm/pachyderm/values.schema.json
+++ b/etc/helm/pachyderm/values.schema.json
@@ -774,7 +774,7 @@
                     "type": "integer"
                 },
                 "goMemLimit": {
-                    "type": "integer"
+                    "type": "string"
                 },
                 "gcPercent": {
                     "type": "integer"