--- a/etc/helm/pachyderm/values.yaml
+++ b/etc/helm/pachyderm/values.yaml
@@ -362,8 +362,14 @@ pachd:
   clusterDeploymentID: ""
   configJob:
     annotations: {}
-  # goMaxProcs is passed as GOMAXPROCS to the pachd container.
+  # goMaxProcs is passed as GOMAXPROCS to the pachd container.  pachd can automatically pick an
+  # optimal GOMAXPROCS from the configured CPU limit, but this overrides it.
   goMaxProcs: 0
+  # goMemLimit is passed as GOMEMLIMIT to the pachd container. pachd can automatically pick an
+  # optimal GOMEMLIMIT from the configured memory request or limit, but this overrides it.
+  goMemLimit: 0
+  # gcPercent sets the initial garbage collection target percentage.
+  gcPercent: 0
   image:
     repository: "pachyderm/pachd"
     pullPolicy: "IfNotPresent"